Visual timer
When watching experiments execute, the display automatically pauses when
the time limit for the particular experiment is reached. The elapsed
time had to be moved down into SimulatorExperiment for this to work.

Modified some of old bash scripts by switching their carriage returns
back to UNIX style so Cygwin can recognize them.

Keep adding files to ignore so that GitHub won't try to commit them.