Fix Two Rooms but break Dangerous Foraging
It turns out that the last update I made ages ago had problems. Adding
Dangerous Foraging broke the Two Rooms domain (and others too). I've
reverted the change, but now Dangerous Foraging is broken. The error has
to do with rangefinder sensors, and how they are set up in
MazeRobotPieSlice. This should be fixed eventually, but the Dangerous
Foraging domain is a bust anyway, and can probably be safely ignored.