Beginnings of adversarial robot
Adding a new experiment with an adversarial robot. The xml file exists,
and the .cs file for the experiment exists. The behavior of the robot is
defined in EnemyRobot, but needs to be worked on (currently just goes
straight forward). I also need to define how the robots interact,
probably by specifying a fitness function that takes this into account.
Since the Adversarial RoomClearingExperiment is new, I had to modify
ExperimentWrapper to accept it.

Lastly, the new refresh.bat file is a batch file that copies
re-compiled/re-built files into the data directly so I can run them
without manually navigating the directories to copy the files ... why
didn't I do this sooner!