EnemyRobot in non-visual mode
It turns out that due to awful code design, the EnemyRobot only worked
in visual mode. Apparently, the list of robots that the enemy was added
to in the adversarial experiment is not used in non-visual mode, only in
visual mode. This has been fixed, so that there should actually be an
enemy present during evolution. Waiting for results...