Robot positioning
Changed the default enemy robot position in the adversarial experiments.
Also made it easier to change this value in the future by tying it to a
parameter specified in the experiment file rather than having it
hard-coded in the source.

The adversarial experiment is now the default experiment loaded, for
convenience.