Enemy tweaking
Changed starting position of enemy, and made it move faster. I don't
know yet whether this makes it too difficult to deal with, but I suspect
there are still ways to handle it.

Behaviorally, I should change the enemy so that it will favor
approaching the evolved bot if it is between itself and a wall (no need
to avoid a wall if target is on the way).