Clamp radians into range
Lots of files changed due to debugging output that is now still present,
but commented out. The only lasting change is in Robot.cs, where any
heading updates, not restrict the heading to the range [0,2PI). It seems
that before, it was completely possible for the heading to go outside of
this range. Surprisingly, this never seemed to cause a problem in the
visual mode, but tracking the robot positions in non-visual mode made it
clear that something weird was happening. Hopefully, the issue had been
fixed now, and both visual and non-visual behavior will be in sync.