Fixing Enemy behavior
Fixing the enemy robot behavior was very frustrating because of slight
differences between the visual and non-visual evaluation modes. The main
problem before was that the EnemyRobot tracked the evolved robot using a
reference that was passed to it at construction. However, in the
non-visual mode, this reference becomes obsolete, so the Enemy always
thinks the evolving bot is at its starting location. I had to make some
changes to assure that the enemy robot always had an up to date
reference to the evolving bot, in order to track it. This lead to many
other changes. Specifically, I had to override the runEnvironment method
of MultiAgentExperiment in my AdversarialRoomClearingExperiment, even
though this method was originally internal to MultiAgentExperiment , and
probably shouldn't have been overridden.

Experiments with an enemy now use a numEnemies variable that is set in
an experiment's xml file.

New survival experiments have been added which are just like the
adversary experiments, except the only fitness is how long the bot
survives. With one enemy only, this is apparently a pretty easy task,
but the important conclusion is that I was able to confirm this by
watching the evolved behavior. More enemies might make the task harder
(or tweaked enemy behavior).

The adversary.plot is a useful gnuplot file that was used in trouble
shooting. Using it requires text to be output from the simulation
tracking the positions of robots. This code is currently commented out,
but still present in the code being submitted.

Another change that makes robot control cleaner, but in retrospect
probably wasn't necessary, is that robot headings are now confined to
the range -PI to PI. Before, any heading was allowed ... this doesn't
seem to cause any problems (though at one point I thought it did), but
it is certainly easier to troubleshoot headings in a fixed range.