Tweaking enemy behavior
The enemy robot still accurately tracks the evolved bot, but careful
observation revealed some cases where it would get stuck in loops or on
corners in stupid ways. Currently, the starting potions of the enemy has
been changed to prevent being stuck in loops (though could maybe still
happen under certain circumstances). The survival experiments have two
enemies instead of one.

The current problem that I am still trying to fix is how to deal with
getting stuck on the corner in the narrow hallway. The evolved bot has
found a position that causes the enemy to be permanently stuck. This is
what to work on next.