More enemy tweaks
Further tweaks to prevent the enemy from getting stuck on walls and
otherwise be unable to catch agent. The starting position change makes
it at least unlikely that two robots will block each other when entering
the narrow hallway.

The change to EXTREME_CLOSE_EVOLVE_THRESHOLD encourages the enemy to
turn into the evolved bot when very close. Note that distance
measurements between bots and walls (via sensors) are a bit hard to
compare because of where the distances are being measured from. Bot
distances seem to be measures from the center of each bot, so the
distance between them can be large (25), even though there are very
close to touching. In contrast, range sensors seem to measure from the
point at which they protrude, so a smaller distance for being too close
to walls makes sense.