Prey Task
Added a domain in which enemy robots are the prey, and the evolved bot
needs to chase them down and collide with them. The EnemyRobot now has a
parameter that can make it run away instead of toward the evolved bot.
The new domain is the prey_two_room experiment, that takes place in the
two room environment. It will probably make sense to mess with the
initial prey robot placement a bit once testing makes it clear that this
domain works as intended.

The prey task fitness is unusual. All fitness apparently has to be
positive, but I want to punish not being able to catch the prey.
Therefore, the agent starts with a large positive fitness that is
subtracted from on each time step that prey are not caught.

Also: The change that made it possible for EnemyRobots to actually track
the evolved bot seems to have slowed down execution ... I also wonder if
perhaps a memory leak has been introduced.