Preference Neuron Links from Separate CPPN Output
Setting dedicatedPreferenceOutputInCPPN to true in an experiment file
that uses preference neuron networks will now create the links to the
preference neurons using a separate output from the CPPN. I've made a
new experiment file for each experiment using a set number of preference
neuron networks that implements this variant ... need to test it out and
see if it helps.

I also made the preference neurons in these substrates have additional
connections to the input layers of networks, which will affect the
original preference neuron approach as well.