Revising patrol fitness functions
In team patrol, I commented out the random +100 bonus.

For lone patrol,
I made the MAX_DISTANCE for regular POIs and the final
goal consistent,
but did not fix the larger issue of the scale
difference with the team
patrol fitness. I did switch from regular distance to Manhattan since
that is used in team patrol