Added the FourTasks domain
This domain is meant to combine all of the previous domains used in the
original MB-HyperNEAT research. The idea is that being able to do all
four domains would be truly multimodal. However, there are still some
issues:

1) Fitness: fitness is currently at wildly different scales, and I'm not
even sure if it all gets added or averaged, but neither makes sense.
Either the fitness scores from the different domains need to be scaled
to the same range, or even better, I need to use multiobjective
optimization. I think that NSGA2 may already be stashed away in this
code somewhere, but I need to look for it.
2) A uniform substrate configuration is now used across all domains.
This means that sensors are present that should be ignored in certain
experiments, but I'm not sure they are behaving correctly. Take a close
look to verify that it all makes sense.