New Robot Model
Fixed problem with robot model. The four tasks domain requires a robot
that supports a signal sensor that is used in the team patrol
environment. However, this sensor is not desired in the Dual Task or Two
Rooms domains. The simple solution (though not great in the long run
because of code duplication) was to create a new robot model that
basically copies the MazeRobotPieSlice model, except it adds a signal
sensor.

Remaining problem: Pie slice sensors activate when they shouldn't.
Specifically, in the team patrol, lone patrol, and dual task hallway
environments. This last one may actually have been a problem in our
original experiments, but not a huge issue.