Fixed Issues with experiment settings
The FourTasks experiment was not working because all environments were
still using the team patrol experiment settings. This was because the
instance of the experiment passed in was always null, due to when the
copy command occurred. This version actually uses different experiment
settings for different environments. However, this indicates that this
code needs to be more thoroughly tested for correctness. I'm still
worried about the network inputs being used in each environment.