Dual task (within four tasks) fix
The fitness for foraging was being used when in the hallway (at least in
visual mode) because the filename includes the full path. Replacing the
name check with EndsWith instead of Equals solves this.

I don't understand why this wouldn't have also been an issue for
previous experiments, but it doesn't seem to have been.