First steps toward multiobjectivity
I've added experiments in the FourTasks domain that use the
multiobjective setting, and I've changed the code to save the objectives
from each environment separately in an array. However, I don't know if
this array is being used the way I think it is, since it is stored
inside of a BehaviorType instance. When executed, this code is clearly
using a multiobjective algorithm, but still outputs results with a focus
on only one objective ... this needs to be changed.