Print individual environment scores
Both the console output and the log output now show scores from the
"objectives" array, which in the case of the FourTasks domain will
contain the individual fitness scores in each environment. Not sure how
this will behave in other domains, but I assume it will only show
behavior metric info.