Non-visual post-eval shows scores from multple objectives
A non-visual post-eval in the FourTasks domain will now show the
individual fitness scores in each environment. This has revealed why the
final evolution scores were not matching the post-eval scores:
Performance in the first domain, the team patrol domain, is worse in the
post-eval than the eval during evolution. This needs to be investigated
further.