XML experiment file fixes
Added settings to the FourTask experiments that are part of the team
patrol experiments. These settings were already in the code (the part
that resets experiment settings) but not in the xml files for some
reason. There were also other minor problems with several experiment
files. However, all of these fixes don't seem to have resolved the issue
with the inconsistent experiment results.