FourTasksFitness always sets Team Patrol parameters
The constructor no longer assumes that the settings from the XML file
are ok for the first run in the first domain. This required a special
case of sending the experiment reference into the fitness function after
the factory constructs it.

However, this still does not make post-eval scores match evolution
scores perfectly in the team patrol domain, so more checks are needed.