Cleanup
Several bits of code cleanup. These changes make it clearer what task is
being used in the FourTasks fitness function, and also add checks to
assure that the right settings are paired with the right task. Redundant
code was also changed, so that the experiment reference is only passed
in once. The fitness function also calls reset any time the task is
changed.

These changes make the code a bit more readable, and slightly effect the
low-level details of how they execute, but the performance outcomes are
still the same, so ultimately this is all just refactoring. This also
means that, unfortunately, the problem of the first team patrol eval
score not matching the evolution score is still present.

Next step is to see if the problem is simply with the first task, or if
it is always with team patrol, no matter the order.