Removed redundant checks
Since the int variable for currentEnvironment indicates what the
environment is, the file name does not need to be checked directly.
Previous testing indicated that the correct files were being used when
expected, so this unnecessary check can be removed moving forward.

Additionally, some testing indicated that the consistency problem is
with whichever task runs first, not simply the team patrol domain. This
makes me wonder if this is a general problem with the simulator that
also effects previous domains (not just FourTasks).