Finally fixed consistency issue!
The problem was that the simulation timestep was set outside of the loop
that went through each environment, so it was not changed for each
environment. As a result, it mattered what the last environment was when
starting a new evaluation, since each had a different timestep. Results
should be consistent now.