Visual vs non-visual fix
For post evals, the visual and non-visual performance was not matching
up. This might still be the case in two rooms, but should be fixed in
all other domains. In particular, the problem before was that the team
patrol fitness was being used in all domains because the experiment
initializes and resets the fitness function whenever an xml file is
loaded (environment or genome). I added a method that is called in the
visualizer to fix this problem whenever it happens.