FourTasks now changes the substrate and robot models for each environment
Rather than have one robot and substrate model common across
environments, the new approach is to actually change the robot and
substrate configuration details for each environment. The needed
substrate files are all loaded at the start, so that the file IO can be
avoided as the CPPN cycles through different substrate configurations.

The whole point of this is to fix the weird fitness issue in lone patrol
(agent with different robot model was able to move faster). I'm not sure
if this fix actually works yet, and it seems that these changes have
reintroduced consistency problems with certain evaluations. I'll need to
explore the issue further, but I think this is still a step in the right
direction, despite the added hassle.