Changed all experiment files for FourTasks
The starting conditions are now set to match the team patrol task. They
no longer use a special substrate and robot model

Additionally, the 1M experiment now claims it uses a preference neuron,
and technically it does have a substrate with a preference neuron.
However, the experiment is also set up so that no new brains will ever
be added, so the experiment is still a test of 1M despite the preference
neuron ... including this helped avoid a weird issue in which the
Khepera robot model treated the fourth (preference) output as indicating
when the robot should stop.

multi-3M was removed since I still need to do a lot of work before
really testing the multiobjective stuff ... I can easily recreate it
when I'm sure these xml files will stop changing so much.