Better reproducability: recreate fitness functions
Whenever the FourTasksFitness function changes the environment, it now
creates a new instance of the sub-fitness function associated with that
environment. This seems to make all post evals identical, except for the
hallway task of the dual task now ... but the score seems close. This is
definitely a step in the right direction, but a little more work is
needed.