Fixed consistency issues?
I think I've figured it all out ... at least on the non-visual side. An
unusual call to setNeuronDensity was actually changing the network to an
incorrect structure in different evaluations in the domain, so this fix
should also encourage better performance. Creating a new instance pack
each time may actually be more than is necessary, but shouldn't hurt.

Just need to evolve and test again ... then even if the non-visual evals
are consistent, I'll need to check the visual ones.