Visualization error fixed
Non-visual post evals are now completely consistent with evolution.
There was an error with visual post-evals, but it has been fixed. Visual
post evals seem to be consistent for each environment except for two
rooms, whose results are similar to evolution results, but not identical
... I think it's close enough that I can consider it fixed. Does the two
rooms domain have noise that the other domains don't?