Renamed TWO ROOMs experiment files
I just realized that most of these files have really confusing names.
For some reason, all of the Module Mutation file names mentioned
twoBrainsPref, which is inaccurate ... this was probably because the MM
files were originally created by copying and modifying the 2M files.

Also, there was confusion about the number 2 appearing in some
filenames. This is meant to distinguish experiments in
"two_room_ENV.xml" vs "two_room_ENV2.xml" which is the harder, more
interesting version that Joel made. Now, the 2 is moved up in the
filename after EXP (vs EXP2) so the distinction is clearer