Last ditch fix of module deletion issue
Apparently, the module deletion issue was still not fixed. Therefore,
instead of relying on crossover not to create networks with 0 output
neurons, there is not a check that returns null when this is the case.
This null result is now interpreted by the evolutionary algorithm as a
signal to skip that attempt to create offspring, and generate a new
random member for the next population.