Track links on post evals
Post eval (non-visual) will not print out the number of CPPN links,
modules, and number of substrate links.

This reveals that the networks do indeed have the number of links
identified in the logs. This means that there is a problem with the
selection process. Maybe the old NSGA-II implementation is screwed up,
or maybe the way that individuals tied for best "actual" fitness get
treated oddly