Always track links in substrate
Now this value will always be calculated for the output logs, though it
is only used as fitness if specifically specified.

However, there does seem to be a problem still with the multiobjective
fitness. The main domain fitness stays the same, but the number of links
increases, which is problematic. I need to fix this.