Maybe fixed multiobjective "best genome" issue?
The sorting operator was using the age of the genome as a tie-breaker in
the sorting. This still should not have mattered since the final
tracking of the best genome was done purely based on fitness, but it
still makes sense to change this.

Keep watching and maybe fix later.