Removed some unnecessary text output that has always bothered me
However, close inspection reveals that it is still possible for the link
count to go up without the fitness changing, which should not be
possible if multiobjective selection is proceeding the way it is
supposed to.

I also noticed the the population size is behaving weirdly. It is not
mu+lambda as with NSGA-II. I think that an archive is accumulating
individuals up to a population size threshold, as with novelty search.

Maybe when several agents are tied for fitness, it is possible for the
wrong one to be removed?