Fixed identification of best network genome?
There were two main problems. The first was that in the
EvolutionAlgorithm, the update of the bestGenome did a check of Fitness
instead of RealFitness. RealFitness is the actual agent fitness.
However, when multiple objectives are used, the Fitness gets swapped
with the NSGA-II Pareto layer rank before sorting at some point, but
doesn't get switched back, causing problems.

I also changed the GenomeComparer so that the second objective (most I
am using) is used as a tie-breaker when the first is tied. However, now
I wonder if this causes problems with NSGA-II sorting. Need to check.