More fixes with best genome sorting
The comparer now makes sure that the Fitness matches the objective[0]
value, which would mean that the current Fitness is the real fitness and
not some meta-fitness for sorting purposes, such as the Pareto rank.

The bestGenome update had a problem. Sometimes the second objective
value would get worse, despite corrections in the sorting with
GenomeComparer. I think this is because although the different
sub-species are sorted, the different species may not be in any order
with respect to each other. Therefore, the first species with the best
fitness has its champion arbitrarily chosen. This has been fixed with
extra conditions.