Lots of work dealing with best genome issue
I don't think the problem is fixed yet (see comment in
EvolutionAlgorithm) but I have made progress. There is still a lot of
debugging output present because I'm going to look at this more
tomorrow.

The problem seems to be that Fitness can either mean RealFitness or Rank
(from multiobjective sorting) depending on context, and this creates
problems when TrimAllSpeciesBackToElite() executes (this is definitely
the command that removes the best individual, maximizing RealFitness and
objective[1] as a tie breaker, from the population)