Another issue with best genome tracking fixed
So, the error was that I wasn't allowing agents with better fitness to
become the best unless they also had better secondary fitness. Now,
primary fitness is most important, and secondary fitness is only a
tie-breaker.

I'm going to remove all of the debugging output and do a proper length
test run to make sure this works.