This debugging indicates where the problem is, but it is not fixed
The EA has references to genomes in both the population and the
individual species. The way the ResetPopulation method works is that it
takes a list of genomes, deletes several species, adds the new genomes
to species, clears the population list, and then adds back to the
population from the remaining species.

Somehow, the population that remains at the end does not contain the
best individuals.