Solved best genome problem for real?
I feel like I've solved the problem again, after lots of
troubleshooting. Debugging output is still in the code, but the main
problem seems to be in the rankGenomes method. This method was assuming
that the last objective was a genotypic novelty objective (I already
changed code elsewhere that was also based on this assumption), and
would actually replace the saved secondary objective with some novelty
calculation. This no longer happens. If this works, I'll remove the
debugging output.