Is selection too restrictive now?
Removed the comment. Technically, evolution doesn't seem to be making
any mistakes. However, it is also not progressing, ie. not improving
with a secondary objective incomparison with single-objective evolution.
This could be a legitamate result, or an error in the code.

More and more, I'm feeling inclined to undo most of my changes and look
at the issue from the start, using the knowledge I've gained from
looking at the code. Something to consider if my test runs continue to
flatline.