Cleaner way of fixing issues, though they aren't fixed yet
This commit resolves a merge conflict with a commit undo, but also
redoes the relevant work of all the undone commits in a cleaner way.
However, it still doesn't work, in that it seems to prevent new genomes
from having a chance. This seems to be caused by a problem in the
GenomeComparer, but other problems emerge without this check. Try to
tweak.