Maybe fixed the best genome issue
UpdateFitnessStats is updated to take a parameter indicating when
Fitness == rank, so that bestGenome will not be updated. Additionally, I
caught a bug in which I was updating based on fittestgenome instead of
candidate.

The nice thing is that this all seems to work with a simplified version
of GenomeComparer. I was able to revert it back to the way it was.

Just need to run a long expriment to make sure evolutionary progress can
still happen.