Random noise based on clock seed
I changed the random number genrator to not specify a random seed, meaning that it is specified by the clock. As a result, there is a different random sequence for each eval. Interestingly, the exact same scores are still common, at least when using only sensor noise (need to check others), but not always the same. For example, across 5 evals, one bad score may be achieved by 3-4 runs, and one good score by 1-2. So, there are differences between evals, but they end up pushing the result toward one of few specific attractors.